7-to_kv.py

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -3,6 +3,6 @@
 from typing import Tuple, Union
 
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """func doc"""
     return (k, v**2)
